pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ankama_launcher_emulator"
version = "0.1.0"
description = ""
authors = [{ name = "Valentin Alix", email = "valentin-alix.850@outlook.fr" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "requests>1.0.0",
    "thrift>=0.20.0,<0.21",
    "psutil>3.0.0",
    "wmi>=1.5.1,<2; sys_platform == 'win32'",
    "pycryptodomex>=3.21.0,<4",
    "frida>=16.6.5,<17",
    "python-dateutil>=2.9.0.post0,<3",
    "mitmproxy>=12.2.1",
    "pysocks>=1.7.1",
    "protobuf==5.28.3",
    "pyqt6-fluent-widgets>=1.11.1",
    "PyQt6-WebEngine>=6.10,<6.11",
    "pyright>=1.1.408",
    "pyinstaller>=6.19.0",
    "pydantic>=2.11.10",
    "rnet>=2.4.0,<3",
    "pyscrypt>=1.6.2,<2",
    "cryptography>=43.0.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0,<25",
    "frida-tools>=13.7.1",
    "pytest>=9.0.3",
]

[tool.pyright]
exclude = [
    "**/.venv",
    "./ankama_launcher_emulator/gen_zaap",
    "**/__pycache__",
    "**/.*",
    "**/node_modules"
]
